Extracted PosJoint class to seperate code and header file.
