+++ CMakeLists.txt
@@ -42,6 +42,7 @@ include_directories(include ${catkin_INCLUDE_DIRS})
 add_library(${PROJECT_NAME} 
   src/steered_wheel_base_controller.cpp 
   src/joint_base.cpp
+  src/pos_joint.cpp
 )
 
 target_link_libraries(${PROJECT_NAME} ${catkin_LIBRARIES})
